Single-file MD node.
