+++ src/gui/models/nodeGraph/enumRegistry.cpp
@@ -7,7 +7,7 @@
 #include "math/averaging.h"
 #include "math/windowFunction.h"
 #include "nodes/gr/gr.h"
-#include "nodes/md/md.h"
+#include "nodes/md.h"
 
 std::map<std::type_index, std::shared_ptr<EnumOptionsModel>> EnumRegistry::options_;
